created base class for title and description for custom widgets
